[master] norme check

+++ libft/string/ft_strmapi.c
@@ -6,7 +6,7 @@
 /*   By: fkante <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/04/11 15:29:42 by fkante            #+#    #+#             */
-/*   Updated: 2020/03/12 12:23:56 by amartino         ###   ########.fr       */
+/*   Updated: 2020/05/18 23:36:25 by francis          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ char	*ft_strmapi(char const *s, char (*f)(uint32_t, char))
 	if (s == NULL || f == NULL)
 		return (NULL);
 	i = 0;
-	len = (uint32_t)(uint32_t)(uint32_t)(uint32_t)(uint32_t)(uint32_t)(uint32_t)(uint32_t)(uint32_t)ft_strlen(s);
+	len = (uint32_t)ft_strlen(s);
 	if (!(tmp = (char *)malloc(len + 1)))
 		return (NULL);
 	while (i < len)
